Fix length of file_size for Telegram update.

+++ src/execute/telegram/jobs/mod.rs
@@ -265,7 +265,7 @@ pub enum CoconutEventJob {
         display_name: String,
         video_url: String,
         thumb_url: String,
-        video_size: i32,
+        video_size: i64,
         duration: i32,
         height: i32,
         width: i32,
